chore: Remove `AccountStorageDeltaBuilder`
